added ability to calculate a^k1
MIPS does not do accurate multiplication on large numbers so a^k1 will
not be totally accurate when a or k1 get large enough.  It gives the
same wrong answer consistantly though, so I don't think this will be a
huge issue.  If anything, it makes the encryption even more secure